refactor: Propsのtypeのパス修正
